Connect product API to PostgreSQL

+++ models/product.go
@@ -1,11 +1,11 @@
 package models
 
 type Product struct {
-	ID       int    `json:"id"`
-	Name     string `json:"name" binding:"required"`
-	Slug     string `json:"slug"`
-	Price    int    `json:"price" binding:"required,gt=0"`
-	Quantity int    `json:"quantity" binding:"required,gte=0"`
+	ID       int     `json:"id"`
+	Name     string  `json:"name" binding:"required"`
+	Slug     string  `json:"slug"`
+	Price    float64 `json:"price" binding:"required,gt=0"`
+	Quantity int     `json:"quantity" binding:"required,gte=0"`
 }
 
 var Products = []Product{
